move streameditor back
