+++ MFCMapi.vcxproj
@@ -660,7 +660,7 @@ copy "$(TargetDir)$(TargetName).pdb" "$(SolutionDir)"
     <ClInclude Include="UI\Dialogs\ContentsTable\RulesDlg.h" />
     <ClInclude Include="UI\Dialogs\propList\SingleMessageDialog.h" />
     <ClInclude Include="UI\Controls\SortList\SingleMAPIPropListCtrl.h" />
-    <ClInclude Include="UI\Dialogs\Editors\propeditor\StreamEditor.h" />
+    <ClInclude Include="UI\Dialogs\Editors\StreamEditor.h" />
     <ClInclude Include="UI\Dialogs\Editors\TagArrayEditor.h" />
     <ClInclude Include="UI\UIFunctions.h" />
     <ClInclude Include="UI\mapiui.h" />
@@ -730,7 +730,7 @@ copy "$(TargetDir)$(TargetName).pdb" "$(SolutionDir)"
     <ClCompile Include="UI\Controls\SortList\SingleMAPIPropListCtrl.cpp" />
     <ClCompile Include="UI\Dialogs\propList\SingleMessageDialog.cpp" />
     <ClCompile Include="UI\Dialogs\propList\SingleRecipientDialog.cpp" />
-    <ClCompile Include="UI\Dialogs\Editors\propeditor\StreamEditor.cpp" />
+    <ClCompile Include="UI\Dialogs\Editors\StreamEditor.cpp" />
     <ClCompile Include="UI\Dialogs\Editors\TagArrayEditor.cpp" />
     <ClCompile Include="UI\UIFunctions.cpp" />
     <ClCompile Include="UI\ViewPane\CheckPane.cpp" />
