+++ MFCMapi.vcxproj.filters
@@ -110,14 +110,14 @@
     <ClInclude Include="UI\Dialogs\Editors\propeditor\PropertyEditor.h">
       <Filter>Source Files\UI\Dialogs\Editors\propeditor</Filter>
     </ClInclude>
-    <ClInclude Include="UI\Dialogs\Editors\propeditor\StreamEditor.h">
-      <Filter>Source Files\UI\Dialogs\Editors\propeditor</Filter>
-    </ClInclude>
     <ClInclude Include="UI\Dialogs\Editors\propeditor\MultiValuePropertyEditor.h">
       <Filter>Source Files\UI\Dialogs\Editors\propeditor</Filter>
     </ClInclude>
     <ClInclude Include="UI\Dialogs\Editors\propeditor\ipropeditor.h">
       <Filter>Source Files\UI\Dialogs\Editors\propeditor</Filter>
+    </ClInclude>
+    <ClInclude Include="UI\Dialogs\Editors\StreamEditor.h">
+      <Filter>Source Files\UI\Dialogs\Editors</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
@@ -183,14 +183,14 @@
     <ClCompile Include="UI\Dialogs\Editors\propeditor\PropertyEditor.cpp">
       <Filter>Source Files\UI\Dialogs\Editors\propeditor</Filter>
     </ClCompile>
-    <ClCompile Include="UI\Dialogs\Editors\propeditor\StreamEditor.cpp">
-      <Filter>Source Files\UI\Dialogs\Editors\propeditor</Filter>
-    </ClCompile>
     <ClCompile Include="UI\Dialogs\Editors\propeditor\MultiValuePropertyEditor.cpp">
       <Filter>Source Files\UI\Dialogs\Editors\propeditor</Filter>
     </ClCompile>
     <ClCompile Include="UI\Dialogs\Editors\propeditor\ipropeditor.cpp">
       <Filter>Source Files\UI\Dialogs\Editors\propeditor</Filter>
+    </ClCompile>
+    <ClCompile Include="UI\Dialogs\Editors\StreamEditor.cpp">
+      <Filter>Source Files\UI\Dialogs\Editors</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
