+++ UI/Controls/SortList/SingleMAPIPropListCtrl.cpp
@@ -18,7 +18,7 @@
 #include <core/mapi/cache/globalCache.h>
 #include <UI/Dialogs/Editors/Editor.h>
 #include <UI/Dialogs/Editors/RestrictEditor.h>
-#include <UI/Dialogs/Editors/propeditor/StreamEditor.h>
+#include <UI/Dialogs/Editors/StreamEditor.h>
 #include <UI/Dialogs/Editors/TagArrayEditor.h>
 #include <UI/Dialogs/Editors/propeditor/ipropeditor.h>
 #include <UI/Dialogs/Editors/PropertyTagEditor.h>
