+++ UI/Dialogs/Editors/StreamEditor.h
